Verify artifacts concurrently for a single graph
Prior to this change, verifications for a single graph happened synchronously
when the artifact was accessed. With this change, all verifications are deferred
until the first artifact is accessed, then happen concurrently.

This significantly improves performance for signature verification which requires
downloading an `asc` file and computing expensive signatures. However, this is
not perfect yet. In particular, if multiple projects are built concurrently it's
likely that all workers will be busy and that verification will happen serially
on a single thread instead of using all workers, effectively blocking each
other.